# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Schedule Service API",
        "description": "API for managing scheduling services, clients, and providers",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/appointments": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get all appointments",
                "description": "Display a listing of the resource.",
                "operationId": "09fd012ab97514481bde57bbeac2699b",
                "responses": {
                    "200": {
                        "description": "List of all appointments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "client_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "provider_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "service_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "scheduled_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2023-06-15T14:30:00Z"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Create a new appointment",
                "description": "Store a newly created resource in storage.",
                "operationId": "15017f04a457b26e2a870c8a011a407a",
                "requestBody": {
                    "description": "Appointment data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_id",
                                    "provider_id",
                                    "service_id",
                                    "scheduled_at"
                                ],
                                "properties": {
                                    "client_id": {
                                        "description": "Client ID",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "provider_id": {
                                        "description": "Provider ID",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "service_id": {
                                        "description": "Service ID",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "scheduled_at": {
                                        "description": "Appointment date and time",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-06-15T14:30:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Appointment created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Appointment created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "client_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "provider_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "service_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "scheduled_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2023-06-15T14:30:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "client_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The client id field is required."
                                                    }
                                                },
                                                "provider_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The provider id field is required."
                                                    }
                                                },
                                                "service_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The service id field is required."
                                                    }
                                                },
                                                "scheduled_at": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The scheduled at must be a date after now."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}